# app-modules/portal-docs/tests/Fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Fixture API", "version": "1.0.0" },
  "servers": [{ "url": "https://fixture.test/api" }],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/posts/{post}": {
      "delete": {
        "summary": "Delete a post",
        "operationId": "posts.destroy",
        "tags": ["Post"],
        "deprecated": true,
        "parameters": [
          { "name": "post", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "204": { "description": "Deleted." } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "A personal access token." },
      "apiKeyAuth": { "type": "apiKey", "in": "header", "name": "x-api-key", "description": "A service key." }
    },
...
  }
}